+++ src/sudo_intercept.c
@@ -90,7 +90,7 @@ static char **
 copy_vector(char * const *src)
 {
     char **copy;
-    int i, len = 0;
+    size_t i, len = 0;
     debug_decl(copy_vector, SUDO_DEBUG_EXEC);
 
     if (src != NULL) {
@@ -141,7 +141,7 @@ resolve_path(const char *cmnd, char *out_cmnd, size_t out_size)
     endp = cp + strlen(cp);
     while (cp < endp) {
 	char *colon = strchr(cp, ':');
-	dirlen = colon ? (colon - cp) : (endp - cp);
+	dirlen = colon ? (int)(colon - cp) : (int)(endp - cp);
 	if (dirlen == 0) {
 	    /* empty PATH component is the same as "." */
 	    len = snprintf(path, sizeof(path), "./%s", cmnd);
@@ -265,12 +265,12 @@ exec_wrapper(const char *cmnd, char * const argv[], char * const envp[],
 
 	    for (argc = 0; argv[argc] != NULL; argc++)
 		continue;
-	    shargv = sudo_mmap_allocarray(argc + 2, sizeof(char *));
+	    shargv = sudo_mmap_allocarray((size_t)argc + 2, sizeof(char *));
 	    if (shargv == NULL)
 		goto bad;
 	    shargv[0] = "sh";
 	    shargv[1] = ncmnd;
-	    memcpy(shargv + 2, nargv + 1, argc * sizeof(char *));
+	    memcpy(shargv + 2, nargv + 1, (size_t)argc * sizeof(char *));
 	    ((sudo_fn_execve_t)fn)(_PATH_SUDO_BSHELL, (char **)shargv, nenvp);
 	    sudo_mmap_free(shargv);
 	}
@@ -311,7 +311,7 @@ execl_wrapper(int type, const char *name, const char *arg, va_list ap)
     while (va_arg(ap2, char *) != NULL)
 	argc++;
     va_end(ap2);
-    argv = sudo_mmap_allocarray(argc + 1, sizeof(char *));
+    argv = sudo_mmap_allocarray((size_t)argc + 1, sizeof(char *));
     if (argv == NULL)
 	debug_return_int(-1);
